move other checks to `validate`
